Move CMS stuff into own folder
