+++ carbonmark/components/pages/Blog/Post/index.tsx
@@ -7,8 +7,8 @@ import { PageHead } from "components/PageHead";
 import PortableTextRenderer from "components/pages/Resources/PortableTextRenderer";
 import { Navigation } from "components/shared/Navigation";
 import { Text } from "components/Text";
+import { Post } from "lib/cms/queries";
 import { urls as carbonmarkUrls } from "lib/constants";
-import { Post } from "lib/queries";
 import Image from "next/legacy/image";
 import Link from "next/link";
 import { useRouter } from "next/router";
