+++ carbonmark/components/pages/Resources/FeaturedArticles/Article.tsx
@@ -1,5 +1,5 @@
 import { Text } from "components/Text";
-import { FeaturedPost } from "lib/queries";
+import { FeaturedPost } from "lib/cms/queries";
 import Image from "next/legacy/image";
 import Link from "next/link";
 import { useRouter } from "next/router";
