+++ carbonmark/lib/cms/queries.ts
@@ -1,5 +1,5 @@
 import type { PortableTextBlock } from "@portabletext/types";
-import { IS_PRODUCTION } from "./constants";
+import { IS_PRODUCTION } from "lib/constants";
 
 /* Posts can have a hideFromProduction field that may be set to true if the author
 wants to publish their posts but dont want it to appear in the production environment.
